Refactor OutputWriter; Correct tests to run for .jar build;
