+++ src/test/resources/sample-config-invalid-1.yaml
@@ -6,6 +6,10 @@ outputLocation:
 savepointsLocation:
   path: dummy/path
 outputSettings:
+  exportTarget: file
+  saveIntoJdbcRaw: true
+  saveIntoJdbcMerged: true
+  fileFormat: csv
   includeColumnNames: bananas #this is not a boolean
   saveAsSingleFile: true
   saveIntoTimestampDirectory: true
