+++ src/test/resources/sample-config-invalid-2.yaml
@@ -6,6 +6,10 @@ outputLocation:
 savepointsLocation:
   path: dummy/path
 outputSettings:
+  exportTarget: file
+  saveIntoJdbcRaw: true
+  saveIntoJdbcMerged: true
+  fileFormat: csv
   includeColumnNames: true
   saveAsSingleFile: true
   saveIntoTimestampDirectory: true
