+++ src/test/resources/sample-config-invalid-3.yaml
@@ -7,6 +7,10 @@ outputLocation:
 savepointsLocation:
   path: dummy/path
 outputSettings:
+  exportTarget: file
+  saveIntoJdbcRaw: true
+  saveIntoJdbcMerged: true
+  fileFormat: csv
   includeColumnNames: true
   saveAsSingleFile: true
   saveIntoTimestampDirectory: true
