+++ src/test/resources/sample-config-invalid-5.yaml
@@ -6,6 +6,10 @@ outputLocation:
 savepointsLocation:
   path: dummy/path
 outputSettings:
+  exportTarget: file
+  saveIntoJdbcRaw: true
+  saveIntoJdbcMerged: true
+  fileFormat: csv
   includeColumnNames: true
   saveAsSingleFile: true
   #saveIntoTimestampDirectory: true #Should default missing boolean to false
